优化rtsp推拉流

+++ src/protocol/rtsp.cpp
@@ -12,7 +12,7 @@
 
 #define RTP_PAYLOAD_TYPE_H264   96
 #define RTP_PAYLOAD_TYPE_AAC    97
-char Rtsp::response[1024]{0};
+char Rtsp::response[2048]{0};
 
 /*static std::map<std::string, Info> flowInfo;*/
 
@@ -49,12 +49,12 @@ int Rtsp::responseData(int status, const std::string &msg) {
 
 int Rtsp::parseRtsp(std::string &packet, const std::string &data) {
     int ret;
+
+    memset(response, 0, 2048);
     std::vector<std::string> list = split(data, "\r\n");
     char method[20]{0};
     char url[50]{0};
     char protocolVersion[20]{0};
-
-    char responseBuffer[2048]{0};
 
 
     int num = sscanf(list.front().c_str(), "%s %s %s\r\n", method, url, protocolVersion);
@@ -72,7 +72,7 @@ int Rtsp::parseRtsp(std::string &packet, const std::string &data) {
 
     if (strcmp(method, "OPTIONS") == 0) {
         /*回复客户端当前可用的方法*/
-        sprintf(responseBuffer,
+        sprintf(response,
                 "RTSP/1.0 200 OK\r\n"
                 "CSeq: %s\r\n"
                 "Date: %s\r\n"
@@ -83,8 +83,8 @@ int Rtsp::parseRtsp(std::string &packet, const std::string &data) {
                 generatorDate().c_str()
         );
 
-        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-                                  static_cast<int>(strlen(responseBuffer)));
+        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(response),
+                                  static_cast<int>(strlen(response)));
         if (ret < 0) {
             fprintf(stderr, "发送数据失败\n");
             return ret;
@@ -219,81 +219,12 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                 audioConfig.c_str(),
                 audioControl.c_str()
         );
-//        /*这个语法是C++ 17的*/
-//        for (auto&[key, value]: flowInfo) {
-//            if (value.dir == "." + urlUtils.getPath()) {
-//                pullSession = value.session;
-//                dir = value.dir;
-//                break;
-//            }
-//        }
-//        if (pullSession.empty()) {
-//            sprintf(responseBuffer,
-//                    "RTSP/1.0 404 找不到流\r\n"
-//                    "CSeq: %s\r\n"
-//                    "Date: %s\r\n"
-//                    "Server: XiaoFeng\r\n"
-//                    "Connection: Close\r\n"
-//                    "\r\n",
-//                    obj["CSeq"].c_str(),
-//                    generatorDate().c_str()
-//            );
-//            ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-//                                      static_cast<int>(strlen(responseBuffer)));
-//            if (ret < 0) {
-//                fprintf(stderr, "发送数据失败\n");
-//                return ret;
-//            }
-//            return -1;
-//        }
-//        SdpInfo &sdpInfo = flowInfo[pullSession].sdp;
-//        /* todo 这里可以根据写入的流来自己生成sdp的一些信息 */
-//        std::string videoFmtp;
-//        std::string videoRtpmap;
-//        std::string audioFmtp;
-//        std::string audioRtpmap;
-//
-//        for (std::map<std::string, std::string> &map: sdpInfo.media) {
-//            if (map["type"] == "video" && map.count("fmtp") && map.count("rtpmap")) {
-//                videoFmtp = map["fmtp"];
-//                videoRtpmap = map["rtpmap"];
-//                videoControl = map["control"];
-//            } else if (map["type"] == "audio" && map.count("fmtp") && map.count("rtpmap")) {
-//                audioFmtp = map["fmtp"];
-//                audioRtpmap = map["rtpmap"];
-//                audioControl = map["control"];
-//            } else {
-//                fprintf(stderr, "没找到对应的fmtp和rtpmap\n");
-//                return -1;
-//            }
-//        }
-//
-//        sprintf(sdp,
-//                "v=0\r\n"
-//                "o=- 9%ld 0 IN IP4 %s\r\n"
-//                "t=0 0\r\n"
-//                "a=control:*\r\n"
-//                "m=video 0 RTP/AVP 96\r\n"
-//                "a=rtpmap:96 %s\r\n"
-//                "a=fmtp:96 %s\r\n"
-//                "a=control:%s\r\n"
-//                "m=audio 0 RTP/AVP 97\r\n"
-//                "a=rtpmap:97 %s\r\n"
-//                "a=fmtp:97 %s\r\n"
-//                "a=control:%s\r\n",
-//                time(nullptr), urlUtils.getDomain().c_str(),
-//                videoRtpmap.c_str(),
-//                videoFmtp.c_str(),
-//                videoControl.c_str(),
-//                audioRtpmap.c_str(),
-//                audioFmtp.c_str(),
-//                audioControl.c_str()
-//        );
+
         /*给每个拉流的生成一个唯一id*/
         uniqueSession = generate_unique_string();
         /*上面这里使用从推流端传过来的信息，在sdpInfo里，然后返回给拉流端*/
         /*%zu 为无符号整型*/
-        sprintf(responseBuffer,
+        sprintf(response,
                 "RTSP/1.0 200 OK\r\n"
                 "CSeq: %s\r\n"
                 "Content-Base: %s\r\n"
@@ -308,42 +239,22 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                 uniqueSession.c_str(),
                 sdp
         );
-        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-                                  static_cast<int>(strlen(responseBuffer)));
+        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(response),
+                                  static_cast<int>(strlen(response)));
         if (ret < 0) {
             fprintf(stderr, "发送数据失败\n");
             return ret;
         }
     } else if (strcmp(method, "ANNOUNCE") == 0) {
+        flag = true;
         dir = "." + urlUtils.getPath();
         if (std::filesystem::exists(dir)) {
             fprintf(stderr, "目录已存在\n");
             responseData(500, "换个推流地址");
             return -1;
         }
-//        for (std::pair<const std::string, Info> &item: flowInfo) {
-//            if (item.second.dir == urlUtils.getPath()) {
-//                sprintf(responseBuffer,
-//                        "RTSP/1.0 500 换个路径\r\n"
-//                        "CSeq: %s\r\n"
-//                        "Date: %s\r\n"
-//                        "Server: XiaoFeng\r\n"
-//                        "\r\n",
-//                        obj["CSeq"].c_str(),
-//                        generatorDate().c_str()
-//                );
-//                ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-//                                          static_cast<int>(strlen(responseBuffer)));
-//                if (ret < 0) {
-//                    fprintf(stderr, "发送数据失败\n");
-//                    return ret;
-//                }
-//                /*这里直接退出，并且关闭连接，也可以继续等待客户端重新发送请求*/
-//                return -1;
-//            }
-//        }
 
-        //  dir = "." + urlUtils.getPath();
+
         if (obj.count("Content-Length") > 0) {
             int size = std::stoi(obj["Content-Length"]);
 
@@ -355,9 +266,6 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                     /*推流上来会生成一个唯一的session*/
                     uniqueSession = generate_unique_string();
 
-                    /* Info info;
-                     info.session = uniqueSession;
-                     info.dir = "." + urlUtils.getPath();*/
                     ret = parseSdp(sdp, info);
                     if (ret < 0) {
                         responseData(500, "解析SDP失败");
@@ -370,9 +278,8 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                             audioControl = media["control"];
                         }
                     }
-                    //flowInfo[uniqueSession] = info;
 
-                    sprintf(responseBuffer,
+                    sprintf(response,
                             "RTSP/1.0 200 OK\r\n"
                             "CSeq: %s\r\n"
                             "Date: %s\r\n"
@@ -383,8 +290,8 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                             generatorDate().c_str(),
                             uniqueSession.c_str()
                     );
-                    ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-                                              static_cast<int>(strlen(responseBuffer)));
+                    ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(response),
+                                              static_cast<int>(strlen(response)));
                     if (ret < 0) {
                         fprintf(stderr, "发送数据失败\n");
                         return ret;
@@ -456,7 +363,7 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
         }
 
 
-        sprintf(responseBuffer,
+        sprintf(response,
                 "RTSP/1.0 200 OK\r\n"
                 "CSeq: %s\r\n"
                 "Date: %s\r\n"
@@ -469,15 +376,15 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                 protocol, type, interleaved,
                 obj["Session"].c_str()
         );
-        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-                                  static_cast<int>(strlen(responseBuffer)));
+        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(response),
+                                  static_cast<int>(strlen(response)));
         if (ret < 0) {
             fprintf(stderr, "发送数据失败\n");
             return ret;
         }
 
     } else if (strcmp(method, "RECORD") == 0) {
-        sprintf(responseBuffer,
+        sprintf(response,
                 "RTSP/1.0 200 OK\r\n"
                 "CSeq: %s\r\n"
                 "Date: %s\r\n"
@@ -489,15 +396,15 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                 obj["Session"].c_str()
         );
 
-        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-                                  static_cast<int>(strlen(responseBuffer)));
+        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(response),
+                                  static_cast<int>(strlen(response)));
         if (ret < 0) {
             fprintf(stderr, "发送数据失败 -> option\n");
             return ret;
         }
 
         /*这里接收客户端传过来的音视频数据，组帧*/
-        ret = receiveData(packet, obj["Session"]);
+        ret = receiveData(packet);
         if (ret < 0) {
             fprintf(stderr, "receiveData 失败\n");
             responseData(500, "接收音视频数据失败");
@@ -506,29 +413,7 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
 
     } else if (strcmp(method, "PLAY") == 0) {
 
-//        int transportStreamPacketNumber = flowInfo[pullSession].transportStreamPacketNumber;
-//
-//        if (transportStreamPacketNumber - 1 < 0) {
-//            fprintf(stderr, "还没准备好，等一会儿在拉流\n");
-//            sprintf(responseBuffer,
-//                    "RTSP/1.0 500 还没准备好，等一会儿在拉流\r\n"
-//                    "CSeq: %s\r\n"
-//                    "Date: %s\r\n"
-//                    "Server: XiaoFeng\r\n"
-//                    "\r\n",
-//                    obj["CSeq"].c_str(),
-//                    generatorDate().c_str()
-//            );
-//            ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-//                                      static_cast<int>(strlen(responseBuffer)));
-//            if (ret < 0) {
-//                fprintf(stderr, "发送数据失败\n");
-//                return ret;
-//            }
-//            return -1;
-//        }
-
-        sprintf(responseBuffer,
+        sprintf(response,
                 "RTSP/1.0 200 OK\r\n"
                 "CSeq: %s\r\n"
                 "Range: npt=0.000-\r\n"
@@ -537,8 +422,8 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                 obj["CSeq"].c_str(),
                 obj["Session"].c_str()
         );
-        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-                                  static_cast<int>(strlen(responseBuffer)));
+        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(response),
+                                  static_cast<int>(strlen(response)));
         if (ret < 0) {
             fprintf(stderr, "发送数据失败\n");
             return ret;
@@ -550,7 +435,7 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
     } else if (strcmp(method, "TEARDOWN") == 0) {
         //结束会话请求，该请求会停止所有媒体流，并释放服务器上的相关会话数据。
         /*这里推流端和拉流端都会发送这个请求，要各自释放各自的资源*/
-        sprintf(responseBuffer,
+        sprintf(response,
                 "RTSP/1.0 200 OK\r\n"
                 "CSeq: %s\r\n"
                 "Date: %s\r\n"
@@ -561,8 +446,8 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                 generatorDate().c_str(),
                 obj["Session"].c_str()
         );
-        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-                                  static_cast<int>(strlen(responseBuffer)));
+        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(response),
+                                  static_cast<int>(strlen(response)));
         if (ret < 0) {
             fprintf(stderr, "发送数据失败\n");
             return ret;
@@ -575,7 +460,7 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
         printf("url = %s\n", url);
         /*保持连接用的*/
         /*在暂停流媒体播放，定期发送GET_PARAMETER作为心跳包维持连接*/
-        sprintf(responseBuffer,
+        sprintf(response,
                 "RTSP/1.0 200 OK\r\n"
                 "CSeq: %s\r\n"
                 "Date: %s\r\n"
@@ -586,8 +471,8 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
                 generatorDate().c_str(),
                 obj["Session"].c_str()
         );
-        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(responseBuffer),
-                                  static_cast<int>(strlen(responseBuffer)));
+        ret = TcpSocket::sendData(clientSocket, reinterpret_cast<uint8_t *>(response),
+                                  static_cast<int>(strlen(response)));
         if (ret < 0) {
             fprintf(stderr, "发送数据失败 -> option\n");
             return ret;
@@ -602,13 +487,11 @@ indexdeltalength=3：表示音频的访问单元索引差值（AU-Index-delta）
 }
 
 
-int Rtsp::receiveData(std::string &packet, const std::string &session) {
+int Rtsp::receiveData(std::string &packet) {
 
 
     int ret;
 
-    // Info &info = flowInfo[session];
-    // RtspReceiveData receive(transportStreamPacketNumber);
     RtspReceiveData receive;
     ret = receive.init(clientSocket, dir, videoChannel, audioChannel);
     if (ret < 0) {
@@ -956,8 +839,9 @@ int Rtsp::parseAACConfig(const std::string &config, SdpInfo &sdpInfo) {
 
 Rtsp::~Rtsp() {
     printf("~Rtsp 析构\n");
-    std::filesystem::remove_all(dir);
-    //  flowInfo.erase(uniqueSession);
+    if (flag) {
+        std::filesystem::remove_all(dir);
+    }
 
 
     stopVideoSendFlag = false;
